場所調整
